Commiting code with cleanup done to the GUI to remove all the old stats files no longer being written.  Also, modifications to the hdf5 stats file to only contain 2 parameters per size bin to define the neighborhood statistics.

+++ Source/AIM/Common/ReconstructionFunc.h
@@ -96,7 +96,7 @@ public:
 	double **grainmoments;
 	double **quat_symm;
 
-	vector<vector<int> > neighborhood;
+	vector<vector<double> > neighborhood;
 	vector<vector<double> > svbovera;
 	vector<vector<double> > svcovera;
 	vector<vector<double> > svcoverb;
